// src/types/graphql.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /**
   * The DateTime scalar type represents date and time as a string in RFC3339 format.
   * For example: "1985-04-12T23:20:50.52Z" represents 20 minutes and 50.52 seconds after the 23rd hour of April 12th, 1985 in UTC.
   */
  DateTime: any;
  /**
   * The Int64 scalar type represents a signed 64‐bit numeric non‐fractional value.
   * Int64 can represent values in range [-(2^63),(2^63 - 1)].
   */
  Int64: any;
};
// ...
export type User = {
  __typename?: 'User';
  username: Scalars['String'];
  displayName?: Maybe<Scalars['String']>;
  image?: Maybe<Scalars['String']>;
  isAdmin?: Maybe<Scalars['Boolean']>;
  tickets?: Maybe<Array<Maybe<Ticket>>>;
  authoredComments?: Maybe<Array<Maybe<Comment>>>;
  ticketsAggregate?: Maybe<TicketAggregateResult>;
  authoredCommentsAggregate?: Maybe<CommentAggregateResult>;
};
// ...
export type Column = {
  __typename?: 'Column';
  colID: Scalars['ID'];
  inProject: Project;
  name: Scalars['String'];
  /** @deprecated To order columns inside projects use Project.order */
  orderPreference?: Maybe<Scalars['Int']>;
  tickets?: Maybe<Array<Maybe<Ticket>>>;
  /** A JSON array storing the order of the tickets by id */
  order?: Maybe<Scalars['String']>;
  ticketsAggregate?: Maybe<TicketAggregateResult>;
};
// ...
export type ColumnAggregateResult = {
  __typename?: 'ColumnAggregateResult';
  count?: Maybe<Scalars['Int']>;
  nameMin?: Maybe<Scalars['String']>;
  nameMax?: Maybe<Scalars['String']>;
  orderPreferenceMin?: Maybe<Scalars['Int']>;
  orderPreferenceMax?: Maybe<Scalars['Int']>;
  orderPreferenceSum?: Maybe<Scalars['Int']>;
  orderPreferenceAvg?: Maybe<Scalars['Float']>;
  orderMin?: Maybe<Scalars['String']>;
  orderMax?: Maybe<Scalars['String']>;
};
// ...
export type Project = {
  __typename?: 'Project';
  projID: Scalars['ID'];
  name: Scalars['String'];
  url?: Maybe<Scalars['String']>;
  ghOwner?: Maybe<Scalars['String']>;
  ghName?: Maybe<Scalars['String']>;
  description?: Maybe<Scalars['String']>;
  admin?: Maybe<User>;
  roles?: Maybe<Array<Maybe<Role>>>;
  columns?: Maybe<Array<Maybe<Column>>>;
  /** A JSON array storing the order of the columns by id */
  order?: Maybe<Scalars['String']>;
  rolesAggregate?: Maybe<RoleAggregateResult>;
  columnsAggregate?: Maybe<ColumnAggregateResult>;
};
// ...
export type Role = {
  __typename?: 'Role';
  id: Scalars['ID'];
  permission: Array<Permission>;
  assignedTo?: Maybe<Array<Maybe<User>>>;
  assignedToAggregate?: Maybe<UserAggregateResult>;
};
// ...
export type RoleAggregateResult = {
  __typename?: 'RoleAggregateResult';
  count?: Maybe<Scalars['Int']>;
};
// ...
export type UserAggregateResult = {
  __typename?: 'UserAggregateResult';
  count?: Maybe<Scalars['Int']>;
  usernameMin?: Maybe<Scalars['String']>;
  usernameMax?: Maybe<Scalars['String']>;
  displayNameMin?: Maybe<Scalars['String']>;
  displayNameMax?: Maybe<Scalars['String']>;
  imageMin?: Maybe<Scalars['String']>;
  imageMax?: Maybe<Scalars['String']>;
};
// ...
export type TicketAggregateResult = {
  __typename?: 'TicketAggregateResult';
  count?: Maybe<Scalars['Int']>;
  titleMin?: Maybe<Scalars['String']>;
  titleMax?: Maybe<Scalars['String']>;
  datetimeMin?: Maybe<Scalars['DateTime']>;
  datetimeMax?: Maybe<Scalars['DateTime']>;
  descriptionMin?: Maybe<Scalars['String']>;
  descriptionMax?: Maybe<Scalars['String']>;
  orderPreferenceMin?: Maybe<Scalars['Int']>;
  orderPreferenceMax?: Maybe<Scalars['Int']>;
  orderPreferenceSum?: Maybe<Scalars['Int']>;
  orderPreferenceAvg?: Maybe<Scalars['Float']>;
};
// ...
export type CommentAggregateResult = {
  __typename?: 'CommentAggregateResult';
  count?: Maybe<Scalars['Int']>;
  datetimeMin?: Maybe<Scalars['DateTime']>;
  datetimeMax?: Maybe<Scalars['DateTime']>;
  textMin?: Maybe<Scalars['String']>;
  textMax?: Maybe<Scalars['String']>;
};
// ...
export enum Permission {
  View = 'VIEW',
  Edit = 'EDIT',
  Delete = 'DELETE'
}
// ...
export type Ticket = {
  __typename?: 'Ticket';
  id: Scalars['ID'];
  onColumn: Column;
  title: Scalars['String'];
  datetime?: Maybe<Scalars['DateTime']>;
  description?: Maybe<Scalars['String']>;
  /** @deprecated To order tickets inside columns use Column.order */
  orderPreference?: Maybe<Scalars['Int']>;
  assigned?: Maybe<User>;
  comments?: Maybe<Array<Maybe<Comment>>>;
  commentsAggregate?: Maybe<CommentAggregateResult>;
};
// ...
export type Comment = {
  __typename?: 'Comment';
  id: Scalars['ID'];
  datetime: Scalars['DateTime'];
  text: Scalars['String'];
  onTicket?: Maybe<Ticket>;
  author: User;
};